Adicionar calculo de dinheiro na tela de preview
