+++ macroChallengeApp/macroChallengeApp/Flows/Current/Extensions/MyTripViewControllerExtension.swift
@@ -116,7 +116,7 @@ extension MyTripViewController: UICollectionViewDelegate {
             updateConstraintsTable()
             self.myTripView.activitiesTableView.reloadData()
             Task {
-                await self.updateBudget()
+                await currencyController.updateBudget(activites: activites, userCurrency: self.userCurrency)
             }
             self.updateTotalBudgetValue()
         }
@@ -251,11 +251,7 @@ extension MyTripViewController: UITableViewDelegate {
         let deleteAction = UIContextualAction(style: .normal,
                                               title: "Delete".localized()) { [weak self] _, _, completionHandler in
             self?.deleteItem(indexPath: indexPath, tableView: tableView)
-            Task {
-                await self!.updateBudget()
-                await self!.updateBudgetTotal()
-                await self!.updateTotalBudgetValue()
-            }
+            self?.updateAllBudget()
             completionHandler(true)
         }
         editAction.backgroundColor = .blueBeach
