+++ macroChallengeApp/macroChallengeApp/Flows/Explore/Extension/ExploreViewController+PreviewRoadmap.swift
@@ -216,7 +216,6 @@ extension PreviewRoadmapViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         guard let cell = tableView.dequeueReusableCell(withIdentifier: ActivityTableViewCell.identifier, for: indexPath) as? ActivityTableViewCell else {
             fatalError("TableCell not found")
-            
         }
         
         let activity = roadmap.days[self.daySelected].activity[indexPath.row]
@@ -225,11 +224,12 @@ extension PreviewRoadmapViewController: UITableViewDataSource {
             cell.localButton.isHidden = true
         }
         cell.localButton.addTarget(self, action: #selector(addRoute(sender:)), for: .touchUpInside)
+        
         cell.setupDaysActivities(hour: activity.hour, currency: activity.currency,
                                  value: String(activity.budget),
                                  name: activity.name)
         cell.activityIcon.image = UIImage(named: activity.category)
-        
+        self.updateAllBudget()
         return cell
     }
